Changed types in MessageContext: Value (uint -> uint64), Time (uint -> int64), Balance (uint -> uint64) and in TransferAction: Value (float64 -> uint64)

+++ cli/gochl.go
@@ -48,7 +48,7 @@ func main() {
 	msgCtx := &chainlog.MessageContext{
 		Sender:  "0x0",
 		Value:   0,
-		Time:    uint(time.Now().Unix()),
+		Time:    time.Now().Unix(),
 		Balance: 0,
 	}
